Sync workspace: fix Epoch field access, duplicate modules, test assertions
Major changes from codebase audit:
- Fix Epoch(u64) private field access (.0 -> .get()) across all crates
- Fix WorkflowLineage private field access (direct -> accessor methods)
- Remove duplicate module files (lineage.rs, fs_store.rs, types.rs, etc.)
- Update test assertions for new CLI commands (compensate, history, serve)
- Add workflow_binary_hash to V3StartRequest across vo-api
- Remove DagNode capability field from command_history constructions
- Replace deprecated TimeoutLayer::new() with with_status_code()
- Replace criterion::black_box with std::hint::black_box in benches
- Add Serialize/Deserialize derives to TimerRecord
- Fix clippy warnings across workspace

+++ crates/vo-storage/src/key_encoding/red_queen_adversarial/prefix_collision.rs
@@ -67,7 +67,7 @@ fn red_queen_dedupe_key_has_length_prefix_format() {
 
     assert_eq!(
         dedupe_prefix.len(),
-        2 + dedupe_str.len() as usize,
+        2 + dedupe_str.len(),
         "BUG: dedupe prefix should be 2-byte length + key bytes"
     );
